指定されたように修正
