+++ frontend/next.config.ts
@@ -3,12 +3,11 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   images: {
     remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "avatars.githubusercontent.com",
-        port: "",
-        pathname: "/**",
-      },
+      { protocol: "https", hostname: "avatars.githubusercontent.com" },
+      { protocol: "https", hostname: "images.unsplash.com" },
+      { protocol: "https", hostname: "zenn.dev" },
+      { protocol: "https", hostname: "cdn.qiita.com" },
+      { protocol: "https", hostname: "assets.st-note.com" },
     ],
   },
 };
